requirements.txt
torch>=2.0
torchvision
diffusers>=0.25.0
transformers>=4.36.0
accelerate
opencv-python
numpy
Pillow
tqdm
controlnet-aux
safetensors
